+++ host/continue_as_new_test.go
