+++ host/workflow_test.go
@@ -1,0 +1,794 @@
+// The MIT License
+//
+// Copyright (c) 2020 Temporal Technologies Inc.  All rights reserved.
+//
+// Copyright (c) 2020 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package host
+
+import (
+	"bytes"
+	"encoding/binary"
+	"math"
+	"strconv"
+	"time"
+
+	"github.com/pborman/uuid"
+	commandpb "go.temporal.io/api/command/v1"
+	commonpb "go.temporal.io/api/common/v1"
+	enumspb "go.temporal.io/api/enums/v1"
+	historypb "go.temporal.io/api/history/v1"
+	"go.temporal.io/api/serviceerror"
+	taskqueuepb "go.temporal.io/api/taskqueue/v1"
+	"go.temporal.io/api/workflowservice/v1"
+
+	"go.temporal.io/server/common"
+	"go.temporal.io/server/common/convert"
+	"go.temporal.io/server/common/failure"
+	"go.temporal.io/server/common/log/tag"
+	"go.temporal.io/server/common/payloads"
+	"go.temporal.io/server/common/primitives/timestamp"
+	"go.temporal.io/server/service/matching"
+)
+
+func (s *integrationSuite) TestStartWorkflowExecution() {
+	id := "integration-start-workflow-test"
+	wt := "integration-start-workflow-test-type"
+	tl := "integration-start-workflow-test-taskqueue"
+	identity := "worker1"
+
+	request := &workflowservice.StartWorkflowExecutionRequest{
+		RequestId:          uuid.New(),
+		Namespace:          s.namespace,
+		WorkflowId:         id,
+		WorkflowType:       &commonpb.WorkflowType{Name: wt},
+		TaskQueue:          &taskqueuepb.TaskQueue{Name: tl},
+		Input:              nil,
+		WorkflowRunTimeout: timestamp.DurationPtr(100 * time.Second),
+		Identity:           identity,
+	}
+
+	we0, err0 := s.engine.StartWorkflowExecution(NewContext(), request)
+	s.NoError(err0)
+
+	// Validate the default value for WorkflowTaskTimeoutSeconds
+	historyResponse, err := s.engine.GetWorkflowExecutionHistory(NewContext(), &workflowservice.GetWorkflowExecutionHistoryRequest{
+		Namespace: s.namespace,
+		Execution: &commonpb.WorkflowExecution{
+			WorkflowId: id,
+			RunId:      we0.RunId,
+		},
+	})
+	s.NoError(err)
+	history := historyResponse.History
+	startedEvent := history.Events[0].GetWorkflowExecutionStartedEventAttributes()
+	s.Equal(10*time.Second, timestamp.DurationValue(startedEvent.GetWorkflowTaskTimeout()))
+
+	we1, err1 := s.engine.StartWorkflowExecution(NewContext(), request)
+	s.NoError(err1)
+	s.Equal(we0.RunId, we1.RunId)
+
+	newRequest := &workflowservice.StartWorkflowExecutionRequest{
+		RequestId:           uuid.New(),
+		Namespace:           s.namespace,
+		WorkflowId:          id,
+		WorkflowType:        &commonpb.WorkflowType{Name: wt},
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Input:               nil,
+		WorkflowRunTimeout:  timestamp.DurationPtr(100 * time.Second),
+		WorkflowTaskTimeout: timestamp.DurationPtr(1 * time.Second),
+		Identity:            identity,
+	}
+	we2, err2 := s.engine.StartWorkflowExecution(NewContext(), newRequest)
+	s.NotNil(err2)
+	s.IsType(&serviceerror.WorkflowExecutionAlreadyStarted{}, err2)
+	s.Nil(we2)
+}
+
+func (s *integrationSuite) TestTerminateWorkflow() {
+	id := "integration-terminate-workflow-test"
+	wt := "integration-terminate-workflow-test-type"
+	tl := "integration-terminate-workflow-test-taskqueue"
+	identity := "worker1"
+	activityName := "activity_type1"
+
+	request := &workflowservice.StartWorkflowExecutionRequest{
+		RequestId:           uuid.New(),
+		Namespace:           s.namespace,
+		WorkflowId:          id,
+		WorkflowType:        &commonpb.WorkflowType{Name: wt},
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Input:               nil,
+		WorkflowRunTimeout:  timestamp.DurationPtr(100 * time.Second),
+		WorkflowTaskTimeout: timestamp.DurationPtr(1 * time.Second),
+		Identity:            identity,
+	}
+
+	we, err0 := s.engine.StartWorkflowExecution(NewContext(), request)
+	s.NoError(err0)
+
+	s.Logger.Info("StartWorkflowExecution", tag.WorkflowRunID(we.RunId))
+
+	activityCount := int32(1)
+	activityCounter := int32(0)
+	wtHandler := func(execution *commonpb.WorkflowExecution, wt *commonpb.WorkflowType,
+		previousStartedEventID, startedEventID int64, history *historypb.History) ([]*commandpb.Command, error) {
+		if activityCounter < activityCount {
+			activityCounter++
+			buf := new(bytes.Buffer)
+			s.Nil(binary.Write(buf, binary.LittleEndian, activityCounter))
+
+			return []*commandpb.Command{{
+				CommandType: enumspb.COMMAND_TYPE_SCHEDULE_ACTIVITY_TASK,
+				Attributes: &commandpb.Command_ScheduleActivityTaskCommandAttributes{ScheduleActivityTaskCommandAttributes: &commandpb.ScheduleActivityTaskCommandAttributes{
+					ActivityId:             convert.Int32ToString(activityCounter),
+					ActivityType:           &commonpb.ActivityType{Name: activityName},
+					TaskQueue:              &taskqueuepb.TaskQueue{Name: tl},
+					Input:                  payloads.EncodeBytes(buf.Bytes()),
+					ScheduleToCloseTimeout: timestamp.DurationPtr(100 * time.Second),
+					ScheduleToStartTimeout: timestamp.DurationPtr(10 * time.Second),
+					StartToCloseTimeout:    timestamp.DurationPtr(50 * time.Second),
+					HeartbeatTimeout:       timestamp.DurationPtr(5 * time.Second),
+				}},
+			}}, nil
+		}
+
+		return []*commandpb.Command{{
+			CommandType: enumspb.COMMAND_TYPE_COMPLETE_WORKFLOW_EXECUTION,
+			Attributes: &commandpb.Command_CompleteWorkflowExecutionCommandAttributes{CompleteWorkflowExecutionCommandAttributes: &commandpb.CompleteWorkflowExecutionCommandAttributes{
+				Result: payloads.EncodeString("Done"),
+			}},
+		}}, nil
+	}
+
+	atHandler := func(execution *commonpb.WorkflowExecution, activityType *commonpb.ActivityType,
+		activityID string, input *commonpb.Payloads, taskToken []byte) (*commonpb.Payloads, bool, error) {
+
+		return payloads.EncodeString("Activity Result"), false, nil
+	}
+
+	poller := &TaskPoller{
+		Engine:              s.engine,
+		Namespace:           s.namespace,
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Identity:            identity,
+		WorkflowTaskHandler: wtHandler,
+		ActivityTaskHandler: atHandler,
+		Logger:              s.Logger,
+		T:                   s.T(),
+	}
+
+	_, err := poller.PollAndProcessWorkflowTask(false, false)
+	s.Logger.Info("PollAndProcessWorkflowTask", tag.Error(err))
+	s.NoError(err)
+
+	terminateReason := "terminate reason"
+	terminateDetails := payloads.EncodeString("terminate details")
+	_, err = s.engine.TerminateWorkflowExecution(NewContext(), &workflowservice.TerminateWorkflowExecutionRequest{
+		Namespace: s.namespace,
+		WorkflowExecution: &commonpb.WorkflowExecution{
+			WorkflowId: id,
+			RunId:      we.RunId,
+		},
+		Reason:   terminateReason,
+		Details:  terminateDetails,
+		Identity: identity,
+	})
+	s.NoError(err)
+
+	executionTerminated := false
+GetHistoryLoop:
+	for i := 0; i < 10; i++ {
+		historyResponse, err := s.engine.GetWorkflowExecutionHistory(NewContext(), &workflowservice.GetWorkflowExecutionHistoryRequest{
+			Namespace: s.namespace,
+			Execution: &commonpb.WorkflowExecution{
+				WorkflowId: id,
+				RunId:      we.RunId,
+			},
+		})
+		s.NoError(err)
+		history := historyResponse.History
+
+		lastEvent := history.Events[len(history.Events)-1]
+		if lastEvent.EventType != enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_TERMINATED {
+			s.Logger.Warn("Execution not terminated yet")
+			time.Sleep(100 * time.Millisecond)
+			continue GetHistoryLoop
+		}
+
+		terminateEventAttributes := lastEvent.GetWorkflowExecutionTerminatedEventAttributes()
+		s.Equal(terminateReason, terminateEventAttributes.Reason)
+		s.Equal(terminateDetails, terminateEventAttributes.Details)
+		s.Equal(identity, terminateEventAttributes.Identity)
+		executionTerminated = true
+		break GetHistoryLoop
+	}
+
+	s.True(executionTerminated)
+
+	newExecutionStarted := false
+StartNewExecutionLoop:
+	for i := 0; i < 10; i++ {
+		request := &workflowservice.StartWorkflowExecutionRequest{
+			RequestId:           uuid.New(),
+			Namespace:           s.namespace,
+			WorkflowId:          id,
+			WorkflowType:        &commonpb.WorkflowType{Name: wt},
+			TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+			Input:               nil,
+			WorkflowRunTimeout:  timestamp.DurationPtr(100 * time.Second),
+			WorkflowTaskTimeout: timestamp.DurationPtr(1 * time.Second),
+			Identity:            identity,
+		}
+
+		newExecution, err := s.engine.StartWorkflowExecution(NewContext(), request)
+		if err != nil {
+			s.Logger.Warn("Start New Execution failed. Error", tag.Error(err))
+			time.Sleep(100 * time.Millisecond)
+			continue StartNewExecutionLoop
+		}
+
+		s.Logger.Info("New Execution Started with the same ID", tag.WorkflowID(id),
+			tag.WorkflowRunID(newExecution.RunId))
+		newExecutionStarted = true
+		break StartNewExecutionLoop
+	}
+
+	s.True(newExecutionStarted)
+}
+
+func (s *integrationSuite) TestSequentialWorkflow() {
+	id := "integration-sequential-workflow-test"
+	wt := "integration-sequential-workflow-test-type"
+	tl := "integration-sequential-workflow-test-taskqueue"
+	identity := "worker1"
+	activityName := "activity_type1"
+
+	request := &workflowservice.StartWorkflowExecutionRequest{
+		RequestId:           uuid.New(),
+		Namespace:           s.namespace,
+		WorkflowId:          id,
+		WorkflowType:        &commonpb.WorkflowType{Name: wt},
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Input:               nil,
+		WorkflowRunTimeout:  timestamp.DurationPtr(100 * time.Second),
+		WorkflowTaskTimeout: timestamp.DurationPtr(1 * time.Second),
+		Identity:            identity,
+	}
+
+	we, err0 := s.engine.StartWorkflowExecution(NewContext(), request)
+	s.NoError(err0)
+
+	s.Logger.Info("StartWorkflowExecution", tag.WorkflowRunID(we.RunId))
+
+	workflowComplete := false
+	activityCount := int32(10)
+	activityCounter := int32(0)
+	wtHandler := func(execution *commonpb.WorkflowExecution, wt *commonpb.WorkflowType,
+		previousStartedEventID, startedEventID int64, history *historypb.History) ([]*commandpb.Command, error) {
+		if activityCounter < activityCount {
+			activityCounter++
+			buf := new(bytes.Buffer)
+			s.Nil(binary.Write(buf, binary.LittleEndian, activityCounter))
+
+			return []*commandpb.Command{{
+				CommandType: enumspb.COMMAND_TYPE_SCHEDULE_ACTIVITY_TASK,
+				Attributes: &commandpb.Command_ScheduleActivityTaskCommandAttributes{ScheduleActivityTaskCommandAttributes: &commandpb.ScheduleActivityTaskCommandAttributes{
+					ActivityId:             convert.Int32ToString(activityCounter),
+					ActivityType:           &commonpb.ActivityType{Name: activityName},
+					TaskQueue:              &taskqueuepb.TaskQueue{Name: tl},
+					Input:                  payloads.EncodeBytes(buf.Bytes()),
+					ScheduleToCloseTimeout: timestamp.DurationPtr(100 * time.Second),
+					ScheduleToStartTimeout: timestamp.DurationPtr(10 * time.Second),
+					StartToCloseTimeout:    timestamp.DurationPtr(50 * time.Second),
+					HeartbeatTimeout:       timestamp.DurationPtr(5 * time.Second),
+				}},
+			}}, nil
+		}
+
+		workflowComplete = true
+		return []*commandpb.Command{{
+			CommandType: enumspb.COMMAND_TYPE_COMPLETE_WORKFLOW_EXECUTION,
+			Attributes: &commandpb.Command_CompleteWorkflowExecutionCommandAttributes{CompleteWorkflowExecutionCommandAttributes: &commandpb.CompleteWorkflowExecutionCommandAttributes{
+				Result: payloads.EncodeString("Done"),
+			}},
+		}}, nil
+	}
+
+	expectedActivity := int32(1)
+	atHandler := func(execution *commonpb.WorkflowExecution, activityType *commonpb.ActivityType,
+		activityID string, input *commonpb.Payloads, taskToken []byte) (*commonpb.Payloads, bool, error) {
+		s.EqualValues(id, execution.WorkflowId)
+		s.Equal(activityName, activityType.Name)
+		id, _ := strconv.Atoi(activityID)
+		s.Equal(int(expectedActivity), id)
+		var b []byte
+		err := payloads.Decode(input, &b)
+		s.NoError(err)
+		buf := bytes.NewReader(b)
+		var in int32
+		binary.Read(buf, binary.LittleEndian, &in)
+		s.Equal(expectedActivity, in)
+		expectedActivity++
+
+		return payloads.EncodeString("Activity Result"), false, nil
+	}
+
+	poller := &TaskPoller{
+		Engine:              s.engine,
+		Namespace:           s.namespace,
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Identity:            identity,
+		WorkflowTaskHandler: wtHandler,
+		ActivityTaskHandler: atHandler,
+		Logger:              s.Logger,
+		T:                   s.T(),
+	}
+
+	for i := 0; i < 10; i++ {
+		_, err := poller.PollAndProcessWorkflowTask(false, false)
+		s.Logger.Info("PollAndProcessWorkflowTask", tag.Error(err))
+		s.NoError(err)
+		if i%2 == 0 {
+			err = poller.PollAndProcessActivityTask(false)
+		} else { // just for testing respondActivityTaskCompleteByID
+			err = poller.PollAndProcessActivityTaskWithID(false)
+		}
+		s.Logger.Info("PollAndProcessActivityTask", tag.Error(err))
+		s.NoError(err)
+	}
+
+	s.False(workflowComplete)
+	_, err := poller.PollAndProcessWorkflowTask(true, false)
+	s.NoError(err)
+	s.True(workflowComplete)
+}
+
+func (s *integrationSuite) TestCompleteWorkflowTaskAndCreateNewOne() {
+	id := "integration-complete-workflow-task-create-new-test"
+	wt := "integration-complete-workflow-task-create-new-test-type"
+	tl := "integration-complete-workflow-task-create-new-test-taskqueue"
+	identity := "worker1"
+
+	request := &workflowservice.StartWorkflowExecutionRequest{
+		RequestId:           uuid.New(),
+		Namespace:           s.namespace,
+		WorkflowId:          id,
+		WorkflowType:        &commonpb.WorkflowType{Name: wt},
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Input:               nil,
+		WorkflowRunTimeout:  timestamp.DurationPtr(100 * time.Second),
+		WorkflowTaskTimeout: timestamp.DurationPtr(1 * time.Second),
+		Identity:            identity,
+	}
+
+	we, err0 := s.engine.StartWorkflowExecution(NewContext(), request)
+	s.NoError(err0)
+
+	s.Logger.Info("StartWorkflowExecution", tag.WorkflowRunID(we.RunId))
+
+	commandCount := 0
+	wtHandler := func(execution *commonpb.WorkflowExecution, wt *commonpb.WorkflowType,
+		previousStartedEventID, startedEventID int64, history *historypb.History) ([]*commandpb.Command, error) {
+
+		if commandCount < 2 {
+			commandCount++
+			return []*commandpb.Command{{
+				CommandType: enumspb.COMMAND_TYPE_RECORD_MARKER,
+				Attributes: &commandpb.Command_RecordMarkerCommandAttributes{RecordMarkerCommandAttributes: &commandpb.RecordMarkerCommandAttributes{
+					MarkerName: "test-marker",
+				}},
+			}}, nil
+		}
+
+		return []*commandpb.Command{{
+			CommandType: enumspb.COMMAND_TYPE_COMPLETE_WORKFLOW_EXECUTION,
+			Attributes: &commandpb.Command_CompleteWorkflowExecutionCommandAttributes{CompleteWorkflowExecutionCommandAttributes: &commandpb.CompleteWorkflowExecutionCommandAttributes{
+				Result: payloads.EncodeString("Done"),
+			}},
+		}}, nil
+	}
+
+	poller := &TaskPoller{
+		Engine:              s.engine,
+		Namespace:           s.namespace,
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		StickyTaskQueue:     &taskqueuepb.TaskQueue{Name: tl},
+		Identity:            identity,
+		WorkflowTaskHandler: wtHandler,
+		Logger:              s.Logger,
+		T:                   s.T(),
+	}
+
+	_, newTask, err := poller.PollAndProcessWorkflowTaskWithAttemptAndRetryAndForceNewWorkflowTask(
+		false,
+		false,
+		true,
+		true,
+		0,
+		1,
+		true,
+		nil)
+	s.NoError(err)
+	s.NotNil(newTask)
+	s.NotNil(newTask.WorkflowTask)
+
+	s.Equal(int64(3), newTask.WorkflowTask.GetPreviousStartedEventId())
+	s.Equal(int64(7), newTask.WorkflowTask.GetStartedEventId())
+	s.Equal(4, len(newTask.WorkflowTask.History.Events))
+	s.Equal(enumspb.EVENT_TYPE_WORKFLOW_TASK_COMPLETED, newTask.WorkflowTask.History.Events[0].GetEventType())
+	s.Equal(enumspb.EVENT_TYPE_MARKER_RECORDED, newTask.WorkflowTask.History.Events[1].GetEventType())
+	s.Equal(enumspb.EVENT_TYPE_WORKFLOW_TASK_SCHEDULED, newTask.WorkflowTask.History.Events[2].GetEventType())
+	s.Equal(enumspb.EVENT_TYPE_WORKFLOW_TASK_STARTED, newTask.WorkflowTask.History.Events[3].GetEventType())
+}
+
+func (s *integrationSuite) TestWorkflowTaskAndActivityTaskTimeoutsWorkflow() {
+	id := "integration-timeouts-workflow-test"
+	wt := "integration-timeouts-workflow-test-type"
+	tl := "integration-timeouts-workflow-test-taskqueue"
+	identity := "worker1"
+	activityName := "activity_timer"
+
+	request := &workflowservice.StartWorkflowExecutionRequest{
+		RequestId:           uuid.New(),
+		Namespace:           s.namespace,
+		WorkflowId:          id,
+		WorkflowType:        &commonpb.WorkflowType{Name: wt},
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Input:               nil,
+		WorkflowRunTimeout:  timestamp.DurationPtr(100 * time.Second),
+		WorkflowTaskTimeout: timestamp.DurationPtr(1 * time.Second),
+		Identity:            identity,
+	}
+
+	we, err0 := s.engine.StartWorkflowExecution(NewContext(), request)
+	s.NoError(err0)
+
+	s.Logger.Info("StartWorkflowExecution", tag.WorkflowRunID(we.RunId))
+
+	workflowComplete := false
+	activityCount := int32(4)
+	activityCounter := int32(0)
+
+	wtHandler := func(execution *commonpb.WorkflowExecution, wt *commonpb.WorkflowType,
+		previousStartedEventID, startedEventID int64, history *historypb.History) ([]*commandpb.Command, error) {
+		if activityCounter < activityCount {
+			activityCounter++
+			buf := new(bytes.Buffer)
+			s.Nil(binary.Write(buf, binary.LittleEndian, activityCounter))
+
+			return []*commandpb.Command{{
+				CommandType: enumspb.COMMAND_TYPE_SCHEDULE_ACTIVITY_TASK,
+				Attributes: &commandpb.Command_ScheduleActivityTaskCommandAttributes{ScheduleActivityTaskCommandAttributes: &commandpb.ScheduleActivityTaskCommandAttributes{
+					ActivityId:             convert.Int32ToString(activityCounter),
+					ActivityType:           &commonpb.ActivityType{Name: activityName},
+					TaskQueue:              &taskqueuepb.TaskQueue{Name: tl},
+					Input:                  payloads.EncodeBytes(buf.Bytes()),
+					ScheduleToCloseTimeout: timestamp.DurationPtr(1 * time.Second),
+					ScheduleToStartTimeout: timestamp.DurationPtr(1 * time.Second),
+					StartToCloseTimeout:    timestamp.DurationPtr(1 * time.Second),
+					HeartbeatTimeout:       timestamp.DurationPtr(1 * time.Second),
+				}},
+			}}, nil
+		}
+
+		s.Logger.Info("Completing enums")
+
+		workflowComplete = true
+		return []*commandpb.Command{{
+			CommandType: enumspb.COMMAND_TYPE_COMPLETE_WORKFLOW_EXECUTION,
+			Attributes: &commandpb.Command_CompleteWorkflowExecutionCommandAttributes{CompleteWorkflowExecutionCommandAttributes: &commandpb.CompleteWorkflowExecutionCommandAttributes{
+				Result: payloads.EncodeString("Done"),
+			}},
+		}}, nil
+	}
+
+	atHandler := func(execution *commonpb.WorkflowExecution, activityType *commonpb.ActivityType,
+		activityID string, input *commonpb.Payloads, taskToken []byte) (*commonpb.Payloads, bool, error) {
+		s.EqualValues(id, execution.WorkflowId)
+		s.Equal(activityName, activityType.Name)
+		s.Logger.Info("Activity ID", tag.WorkflowActivityID(activityID))
+		return payloads.EncodeString("Activity Result"), false, nil
+	}
+
+	poller := &TaskPoller{
+		Engine:              s.engine,
+		Namespace:           s.namespace,
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Identity:            identity,
+		WorkflowTaskHandler: wtHandler,
+		ActivityTaskHandler: atHandler,
+		Logger:              s.Logger,
+		T:                   s.T(),
+	}
+
+	for i := 0; i < 8; i++ {
+		dropWorkflowTask := (i%2 == 0)
+		s.Logger.Info("Calling Workflow Task", tag.Counter(i))
+		var err error
+		if dropWorkflowTask {
+			_, err = poller.PollAndProcessWorkflowTask(true, true)
+		} else {
+			_, err = poller.PollAndProcessWorkflowTaskWithAttempt(true, false, false, false, 2)
+		}
+		if err != nil {
+			historyResponse, err := s.engine.GetWorkflowExecutionHistory(NewContext(), &workflowservice.GetWorkflowExecutionHistoryRequest{
+				Namespace: s.namespace,
+				Execution: &commonpb.WorkflowExecution{
+					WorkflowId: id,
+					RunId:      we.RunId,
+				},
+			})
+			s.NoError(err)
+			history := historyResponse.History
+			common.PrettyPrintHistory(history, s.Logger)
+		}
+		s.True(err == nil || err == matching.ErrNoTasks, "%v", err)
+		if !dropWorkflowTask {
+			s.Logger.Info("Calling PollAndProcessActivityTask", tag.Counter(i))
+			err = poller.PollAndProcessActivityTask(i%4 == 0)
+			s.True(err == nil || err == matching.ErrNoTasks)
+		}
+	}
+
+	s.Logger.Info("Waiting for workflow to complete", tag.WorkflowRunID(we.RunId))
+
+	s.False(workflowComplete)
+	_, err := poller.PollAndProcessWorkflowTask(true, false)
+	s.NoError(err)
+	s.True(workflowComplete)
+}
+
+func (s *integrationSuite) TestWorkflowRetry() {
+	id := "integration-wf-retry-test"
+	wt := "integration-wf-retry-type"
+	tl := "integration-wf-retry-taskqueue"
+	identity := "worker1"
+
+	initialInterval := 1 * time.Second
+	backoffCoefficient := 1.5
+	maximumAttempts := 5
+	request := &workflowservice.StartWorkflowExecutionRequest{
+		RequestId:           uuid.New(),
+		Namespace:           s.namespace,
+		WorkflowId:          id,
+		WorkflowType:        &commonpb.WorkflowType{Name: wt},
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Input:               nil,
+		WorkflowRunTimeout:  timestamp.DurationPtr(100 * time.Second),
+		WorkflowTaskTimeout: timestamp.DurationPtr(1 * time.Second),
+		Identity:            identity,
+		RetryPolicy: &commonpb.RetryPolicy{
+			// Intentionally test server-initialization of Initial Interval value (which should be 1 second)
+			MaximumAttempts:        int32(maximumAttempts),
+			MaximumInterval:        timestamp.DurationPtr(1 * time.Second),
+			NonRetryableErrorTypes: []string{"bad-bug"},
+			BackoffCoefficient:     backoffCoefficient,
+		},
+	}
+
+	we, err0 := s.engine.StartWorkflowExecution(NewContext(), request)
+	s.NoError(err0)
+
+	s.Logger.Info("StartWorkflowExecution", tag.WorkflowRunID(we.RunId))
+
+	var executions []*commonpb.WorkflowExecution
+
+	attemptCount := 1
+
+	wtHandler := func(execution *commonpb.WorkflowExecution, wt *commonpb.WorkflowType,
+		previousStartedEventID, startedEventID int64, history *historypb.History) ([]*commandpb.Command, error) {
+		executions = append(executions, execution)
+		attemptCount++
+		if attemptCount > maximumAttempts {
+			return []*commandpb.Command{
+				{
+					CommandType: enumspb.COMMAND_TYPE_COMPLETE_WORKFLOW_EXECUTION,
+					Attributes: &commandpb.Command_CompleteWorkflowExecutionCommandAttributes{CompleteWorkflowExecutionCommandAttributes: &commandpb.CompleteWorkflowExecutionCommandAttributes{
+						Result: payloads.EncodeString("succeed-after-retry"),
+					}},
+				}}, nil
+		}
+		return []*commandpb.Command{
+			{
+				CommandType: enumspb.COMMAND_TYPE_FAIL_WORKFLOW_EXECUTION,
+				Attributes: &commandpb.Command_FailWorkflowExecutionCommandAttributes{FailWorkflowExecutionCommandAttributes: &commandpb.FailWorkflowExecutionCommandAttributes{
+					Failure: failure.NewServerFailure("retryable-error", false),
+				}},
+			}}, nil
+	}
+
+	poller := &TaskPoller{
+		Engine:              s.engine,
+		Namespace:           s.namespace,
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Identity:            identity,
+		WorkflowTaskHandler: wtHandler,
+		Logger:              s.Logger,
+		T:                   s.T(),
+	}
+
+	describeWorkflowExecution := func(execution *commonpb.WorkflowExecution) (*workflowservice.DescribeWorkflowExecutionResponse, error) {
+		return s.engine.DescribeWorkflowExecution(NewContext(), &workflowservice.DescribeWorkflowExecutionRequest{
+			Namespace: s.namespace,
+			Execution: execution,
+		})
+	}
+
+	for i := 1; i <= maximumAttempts; i++ {
+		_, err := poller.PollAndProcessWorkflowTask(false, false)
+		s.True(err == nil, err)
+		events := s.getHistory(s.namespace, executions[i-1])
+		if i == maximumAttempts {
+			s.Equal(enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_COMPLETED, events[len(events)-1].GetEventType())
+		} else {
+			s.Equal(enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_CONTINUED_AS_NEW, events[len(events)-1].GetEventType())
+		}
+		s.Equal(int32(i), events[0].GetWorkflowExecutionStartedEventAttributes().GetAttempt())
+
+		dweResponse, err := describeWorkflowExecution(executions[i-1])
+		s.NoError(err)
+		backoff := time.Duration(0)
+		if i > 1 {
+			backoff = time.Duration(initialInterval.Seconds()*math.Pow(backoffCoefficient, float64(i-2))) * time.Second
+			// retry backoff cannot larger than MaximumIntervalInSeconds
+			if backoff > time.Second {
+				backoff = time.Second
+			}
+		}
+		expectedExecutionTime := dweResponse.WorkflowExecutionInfo.GetStartTime().Add(backoff)
+		s.Equal(expectedExecutionTime, timestamp.TimeValue(dweResponse.WorkflowExecutionInfo.GetExecutionTime()))
+	}
+}
+
+func (s *integrationSuite) TestWorkflowRetryFailures() {
+	id := "integration-wf-retry-failures-test"
+	wt := "integration-wf-retry-failures-type"
+	tl := "integration-wf-retry-failures-taskqueue"
+	identity := "worker1"
+
+	workflowImpl := func(attempts int, errorReason string, nonRetryable bool, executions *[]*commonpb.WorkflowExecution) workflowTaskHandler {
+		attemptCount := 1
+
+		wtHandler := func(execution *commonpb.WorkflowExecution, wt *commonpb.WorkflowType,
+			previousStartedEventID, startedEventID int64, history *historypb.History) ([]*commandpb.Command, error) {
+			*executions = append(*executions, execution)
+			attemptCount++
+			if attemptCount > attempts {
+				return []*commandpb.Command{
+					{
+						CommandType: enumspb.COMMAND_TYPE_COMPLETE_WORKFLOW_EXECUTION,
+						Attributes: &commandpb.Command_CompleteWorkflowExecutionCommandAttributes{CompleteWorkflowExecutionCommandAttributes: &commandpb.CompleteWorkflowExecutionCommandAttributes{
+							Result: payloads.EncodeString("succeed-after-retry"),
+						}},
+					}}, nil
+			}
+			return []*commandpb.Command{
+				{
+					CommandType: enumspb.COMMAND_TYPE_FAIL_WORKFLOW_EXECUTION,
+					Attributes: &commandpb.Command_FailWorkflowExecutionCommandAttributes{FailWorkflowExecutionCommandAttributes: &commandpb.FailWorkflowExecutionCommandAttributes{
+						// Reason:  "retryable-error",
+						Failure: failure.NewServerFailure(errorReason, nonRetryable),
+					}},
+				}}, nil
+		}
+
+		return wtHandler
+	}
+
+	// Fail using attempt
+	request := &workflowservice.StartWorkflowExecutionRequest{
+		RequestId:           uuid.New(),
+		Namespace:           s.namespace,
+		WorkflowId:          id,
+		WorkflowType:        &commonpb.WorkflowType{Name: wt},
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Input:               nil,
+		WorkflowRunTimeout:  timestamp.DurationPtr(100 * time.Second),
+		WorkflowTaskTimeout: timestamp.DurationPtr(1 * time.Second),
+		Identity:            identity,
+		RetryPolicy: &commonpb.RetryPolicy{
+			InitialInterval:        timestamp.DurationPtr(1 * time.Second),
+			MaximumAttempts:        3,
+			MaximumInterval:        timestamp.DurationPtr(1 * time.Second),
+			NonRetryableErrorTypes: []string{"bad-bug"},
+			BackoffCoefficient:     1,
+		},
+	}
+
+	we, err0 := s.engine.StartWorkflowExecution(NewContext(), request)
+	s.NoError(err0)
+
+	s.Logger.Info("StartWorkflowExecution", tag.WorkflowRunID(we.RunId))
+
+	var executions []*commonpb.WorkflowExecution
+	wtHandler := workflowImpl(5, "retryable-error", false, &executions)
+	poller := &TaskPoller{
+		Engine:              s.engine,
+		Namespace:           s.namespace,
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Identity:            identity,
+		WorkflowTaskHandler: wtHandler,
+		Logger:              s.Logger,
+		T:                   s.T(),
+	}
+
+	_, err := poller.PollAndProcessWorkflowTask(false, false)
+	s.True(err == nil, err)
+	events := s.getHistory(s.namespace, executions[0])
+	s.Equal(enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_CONTINUED_AS_NEW, events[len(events)-1].GetEventType())
+	s.Equal(int32(1), events[0].GetWorkflowExecutionStartedEventAttributes().GetAttempt())
+
+	_, err = poller.PollAndProcessWorkflowTask(false, false)
+	s.True(err == nil, err)
+	events = s.getHistory(s.namespace, executions[1])
+	s.Equal(enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_CONTINUED_AS_NEW, events[len(events)-1].GetEventType())
+	s.Equal(int32(2), events[0].GetWorkflowExecutionStartedEventAttributes().GetAttempt())
+
+	_, err = poller.PollAndProcessWorkflowTask(false, false)
+	s.True(err == nil, err)
+	events = s.getHistory(s.namespace, executions[2])
+	s.Equal(enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_FAILED, events[len(events)-1].GetEventType())
+	s.Equal(int32(3), events[0].GetWorkflowExecutionStartedEventAttributes().GetAttempt())
+
+	// Fail error reason
+	request = &workflowservice.StartWorkflowExecutionRequest{
+		RequestId:           uuid.New(),
+		Namespace:           s.namespace,
+		WorkflowId:          id,
+		WorkflowType:        &commonpb.WorkflowType{Name: wt},
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Input:               nil,
+		WorkflowRunTimeout:  timestamp.DurationPtr(100 * time.Second),
+		WorkflowTaskTimeout: timestamp.DurationPtr(1 * time.Second),
+		Identity:            identity,
+		RetryPolicy: &commonpb.RetryPolicy{
+			InitialInterval:        timestamp.DurationPtr(1 * time.Second),
+			MaximumAttempts:        3,
+			MaximumInterval:        timestamp.DurationPtr(1 * time.Second),
+			NonRetryableErrorTypes: []string{"bad-bug"},
+			BackoffCoefficient:     1,
+		},
+	}
+
+	we, err0 = s.engine.StartWorkflowExecution(NewContext(), request)
+	s.NoError(err0)
+
+	s.Logger.Info("StartWorkflowExecution", tag.WorkflowRunID(we.RunId))
+
+	executions = []*commonpb.WorkflowExecution{}
+	wtHandler = workflowImpl(5, "bad-bug", true, &executions)
+	poller = &TaskPoller{
+		Engine:              s.engine,
+		Namespace:           s.namespace,
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Identity:            identity,
+		WorkflowTaskHandler: wtHandler,
+		Logger:              s.Logger,
+		T:                   s.T(),
+	}
+
+	_, err = poller.PollAndProcessWorkflowTask(false, false)
+	s.True(err == nil, err)
+	events = s.getHistory(s.namespace, executions[0])
+	s.Equal(enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_FAILED, events[len(events)-1].GetEventType())
+	s.Equal(int32(1), events[0].GetWorkflowExecutionStartedEventAttributes().GetAttempt())
+}
