+++ host/workflow_visibility_test.go
@@ -1,0 +1,171 @@
+// The MIT License
+//
+// Copyright (c) 2020 Temporal Technologies Inc.  All rights reserved.
+//
+// Copyright (c) 2020 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package host
+
+import (
+	"time"
+
+	"github.com/pborman/uuid"
+	commandpb "go.temporal.io/api/command/v1"
+	commonpb "go.temporal.io/api/common/v1"
+	enumspb "go.temporal.io/api/enums/v1"
+	filterpb "go.temporal.io/api/filter/v1"
+	historypb "go.temporal.io/api/history/v1"
+	taskqueuepb "go.temporal.io/api/taskqueue/v1"
+	"go.temporal.io/api/workflowservice/v1"
+
+	"go.temporal.io/server/common/payloads"
+	"go.temporal.io/server/common/primitives/timestamp"
+)
+
+func (s *integrationSuite) TestVisibility() {
+	startTime := time.Now().UTC()
+
+	// Start 2 workflow executions
+	id1 := "integration-visibility-test1"
+	id2 := "integration-visibility-test2"
+	wt := "integration-visibility-test-type"
+	tl := "integration-visibility-test-taskqueue"
+	identity := "worker1"
+
+	startRequest := &workflowservice.StartWorkflowExecutionRequest{
+		RequestId:           uuid.New(),
+		Namespace:           s.namespace,
+		WorkflowId:          id1,
+		WorkflowType:        &commonpb.WorkflowType{Name: wt},
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Input:               nil,
+		WorkflowRunTimeout:  timestamp.DurationPtr(100 * time.Second),
+		WorkflowTaskTimeout: timestamp.DurationPtr(10 * time.Second),
+		Identity:            identity,
+	}
+
+	startResponse, err0 := s.engine.StartWorkflowExecution(NewContext(), startRequest)
+	s.NoError(err0)
+
+	// Now complete one of the executions
+	wtHandler := func(execution *commonpb.WorkflowExecution, wt *commonpb.WorkflowType,
+		previousStartedEventID, startedEventID int64, history *historypb.History) ([]*commandpb.Command, error) {
+		return []*commandpb.Command{{
+			CommandType: enumspb.COMMAND_TYPE_COMPLETE_WORKFLOW_EXECUTION,
+			Attributes: &commandpb.Command_CompleteWorkflowExecutionCommandAttributes{CompleteWorkflowExecutionCommandAttributes: &commandpb.CompleteWorkflowExecutionCommandAttributes{
+				Result: payloads.EncodeString("Done"),
+			}},
+		}}, nil
+	}
+
+	poller := &TaskPoller{
+		Engine:              s.engine,
+		Namespace:           s.namespace,
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Identity:            identity,
+		WorkflowTaskHandler: wtHandler,
+		ActivityTaskHandler: nil,
+		Logger:              s.Logger,
+		T:                   s.T(),
+	}
+
+	_, err1 := poller.PollAndProcessWorkflowTask(false, false)
+	s.NoError(err1)
+
+	// wait until the start workflow is done
+	var nextToken []byte
+	historyEventFilterType := enumspb.HISTORY_EVENT_FILTER_TYPE_CLOSE_EVENT
+	for {
+		historyResponse, historyErr := s.engine.GetWorkflowExecutionHistory(NewContext(), &workflowservice.GetWorkflowExecutionHistoryRequest{
+			Namespace: startRequest.Namespace,
+			Execution: &commonpb.WorkflowExecution{
+				WorkflowId: startRequest.WorkflowId,
+				RunId:      startResponse.RunId,
+			},
+			WaitNewEvent:           true,
+			HistoryEventFilterType: historyEventFilterType,
+			NextPageToken:          nextToken,
+		})
+		s.Nil(historyErr)
+		if len(historyResponse.NextPageToken) == 0 {
+			break
+		}
+
+		nextToken = historyResponse.NextPageToken
+	}
+
+	startRequest = &workflowservice.StartWorkflowExecutionRequest{
+		RequestId:           uuid.New(),
+		Namespace:           s.namespace,
+		WorkflowId:          id2,
+		WorkflowType:        &commonpb.WorkflowType{Name: wt},
+		TaskQueue:           &taskqueuepb.TaskQueue{Name: tl},
+		Input:               nil,
+		WorkflowRunTimeout:  timestamp.DurationPtr(100 * time.Second),
+		WorkflowTaskTimeout: timestamp.DurationPtr(10 * time.Second),
+		Identity:            identity,
+	}
+
+	_, err2 := s.engine.StartWorkflowExecution(NewContext(), startRequest)
+	s.NoError(err2)
+
+	startFilter := &filterpb.StartTimeFilter{}
+	startFilter.EarliestTime = &startTime
+	startFilter.LatestTime = timestamp.TimePtr(time.Now().UTC())
+
+	closedCount := 0
+	openCount := 0
+
+	var historyLength int64
+	for i := 0; i < 10; i++ {
+		resp, err3 := s.engine.ListClosedWorkflowExecutions(NewContext(), &workflowservice.ListClosedWorkflowExecutionsRequest{
+			Namespace:       s.namespace,
+			MaximumPageSize: 100,
+			StartTimeFilter: startFilter,
+		})
+		s.NoError(err3)
+		closedCount = len(resp.Executions)
+		if closedCount == 1 {
+			historyLength = resp.Executions[0].HistoryLength
+			break
+		}
+		s.Logger.Info("Closed WorkflowExecution is not yet visible")
+		time.Sleep(100 * time.Millisecond)
+	}
+	s.Equal(1, closedCount)
+	s.Equal(int64(5), historyLength)
+
+	for i := 0; i < 10; i++ {
+		resp, err4 := s.engine.ListOpenWorkflowExecutions(NewContext(), &workflowservice.ListOpenWorkflowExecutionsRequest{
+			Namespace:       s.namespace,
+			MaximumPageSize: 100,
+			StartTimeFilter: startFilter,
+		})
+		s.NoError(err4)
+		openCount = len(resp.Executions)
+		if openCount == 1 {
+			break
+		}
+		s.Logger.Info("Open WorkflowExecution is not yet visible")
+		time.Sleep(100 * time.Millisecond)
+	}
+	s.Equal(1, openCount)
+}
